+++ Integrations/PaloAltoNetworks_PAN_OS_EDL_Management/PaloAltoNetworks_PAN_OS_EDL_Management.py
@@ -104,11 +104,11 @@ def scp_execute(file_name: str, file_path: str):
 
     if SCP_EXTRA_PARAMS:
         param_list = ['scp', '-o', 'StrictHostKeyChecking=no', '-i', CERTIFICATE_FILE.name] + SCP_EXTRA_PARAMS + [
-            file_name, USERNAME + '@' + HOSTNAME + ':' + file_path]
+            file_name, USERNAME + '@' + HOSTNAME + ':' + f'\'{file_path}\'']
         result = subprocess.run(param_list, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
     else:
         param_list = ['scp', '-o', 'StrictHostKeyChecking=no', '-i', CERTIFICATE_FILE.name, file_name,
-                      USERNAME + '@' + HOSTNAME + ':' + file_path]
+                      USERNAME + '@' + HOSTNAME + ':' + f'\'{file_path}\'']
         result = subprocess.run(param_list, stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True)
 
     if result.returncode != 0:
@@ -151,7 +151,7 @@ def edl_get_external_file_command():
 
 
 def edl_search_external_file(file_path: str, search_string: str):
-    return ssh_execute(f'grep {search_string} {file_path}')
+    return ssh_execute(f'grep \'{search_string}\' \'{file_path}\'')
 
 
 def edl_search_external_file_command():
@@ -193,7 +193,7 @@ def edl_update_external_file(file_path: str, list_name: str, verbose: bool):
         return_error('External file was not updated successfully.')
     else:
         if verbose:
-            external_file_items = ssh_execute(f'cat {file_path}')
+            external_file_items = ssh_execute(f'cat \'{file_path}\'')
             if external_file_items:
                 md = tableToMarkdown('Updated File Data:', external_file_items, headers=['Data'])
             else:
@@ -342,7 +342,7 @@ def edl_update_from_external_file_command():
 
 
 def edl_delete_external_file(file_path: str):
-    ssh_execute('rm -f ' + file_path)
+    ssh_execute(f'rm -f \'{file_path}\'')
     return 'File deleted successfully'
 
 
@@ -534,7 +534,7 @@ def edl_get_external_file_metadata_command():
     if DOCUMENT_ROOT:
         file_path = os.path.join(DOCUMENT_ROOT, file_path)
 
-    result = ssh_execute(f'stat {file_path}')
+    result = ssh_execute(f'stat \'{file_path}\'')
 
     file_size = int(result.split("Size: ", 1)[1].split(" ", 1)[0])
     file_name = file_path.split("/")[-1]
@@ -543,7 +543,7 @@ def edl_get_external_file_metadata_command():
     last_modified_parts = result.split("Change: ", 1)[1].split(" ", 2)[0:2]
     last_modified = ' '.join(last_modified_parts)
 
-    number_of_lines = int(ssh_execute(f'wc -l < {file_path}')) + 1
+    number_of_lines = int(ssh_execute(f'wc -l < \'{file_path}\'')) + 1
 
     metadata_outputs = {
         'FileName': file_name,