+++ Integrations/PaloAltoNetworks_PAN_OS_EDL_Management/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+Added the ***pan-os-edl-update-internal-list***, ***pan-os-edl-update-external-file*** commands.
 
 ## [19.8.0] - 2019-08-06
   - Added the ***pan-os-edl-get-external-file-metadata*** command.