+++ Integrations/PaloAltoNetworks_PAN_OS_EDL_Management/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [Unreleased]
-- Fixed: !pan-os-edl-update does't work with file path including space charactors at scp_execute().
-- Fixed: ssh_execute() not working with file name including space characters
+  - Fixed an issue where the ***pan-os-edl-update*** command failed when the file path included space characters at *scp_execute()*.
+  - Fixed an issue where the *ssh_execute()* functioned failed when the file name included space characters.
 
 ## [19.8.0] - 2019-08-06
   - Added the ***pan-os-edl-get-external-file-metadata*** command.