+++ Integrations/PaloAltoNetworks_PAN_OS_EDL_Management/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 
 
+## [19.11.0] - 2019-11-12
+  - Fixed an issue where the ***pan-os-edl-update*** command failed when the file path included space characters at *scp_execute()*.
+  - Fixed an issue where the *ssh_execute()* functioned failed when the file name included space characters.
+
 ## [19.8.0] - 2019-08-06
   - Added the ***pan-os-edl-get-external-file-metadata*** command.
   - When a non-existent list is specified in the ***pan-os-edl-update-from-external-file*** command, the list is automatically created, and the file data is saved to the list.